Add unified signer resolution and registry contract access for standardized queries and transactions.
- Add `resolveSigner()` (`src/utils/signer.ts`) — single entry point that converges `--suri //Alice` dev signers and QR mobile sessions onto `PolkadotSigner`
- Add `getRegistryContract()` (`src/utils/registry.ts`) — returns a typed registry contract handle wired with bundled `cdm.json` and resolved signer
- Add `cdm.json` from `cdm i @example/playground-registry`
- Rename `signer.ts` → `session-signer-patch.ts` to clarify it's a temporary workaround for `@polkadot-apps/terminal`
- Use `parseAsync().then(() => process.exit(0))` in entry point so all commands exit cleanly
