+++ package.json
@@ -18,6 +18,7 @@
         "test:watch": "vitest"
     },
     "dependencies": {
+        "@polkadot-api/sdk-ink": "^0.6.2",
         "@polkadot-apps/address": "latest",
         "@polkadot-apps/bulletin": "latest",
         "@polkadot-apps/chain-client": "latest",
@@ -26,11 +27,10 @@
         "@polkadot-apps/terminal": "latest",
         "@polkadot-apps/tx": "latest",
         "@polkadot-apps/utils": "latest",
-        "@polkadot-api/sdk-ink": "^0.6.2",
         "bulletin-deploy": "latest",
         "commander": "^12.0.0",
-        "polkadot-api": "^1.23.3",
         "ink": "^5.2.1",
+        "polkadot-api": "^1.23.3",
         "react": "^18.3.1",
         "react-devtools-core": "file:stubs/react-devtools-core"
     },
