+++ tsconfig.json
@@ -14,5 +14,8 @@
         "isolatedModules": true,
         "jsx": "react-jsx"
     },
-    "include": ["src"]
+    "include": [
+        "src",
+        "./.cdm/**/*"
+    ]
 }
